# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Vedetta NVR API
  version: "0.1.0"
  description: API for managing cameras, events, recordings, and object detection in the Vedetta Network Video Recorder.
servers:
  - url: /
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/people/{id}:
    get:
      operationId: getPerson
      summary: Get person detail with recent faces
      tags: [people]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Person detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDetail"
        "404":
          description: Person not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updatePerson
      summary: Update a person's name or ignore status
      tags: [people]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonRequest"
      responses:
        "200":
          description: Person updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusUpdated"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deletePerson
      summary: Delete a person
      tags: [people]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Person deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDeleted"
        "400":
          description: Invalid person ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/people/{id}/faces:
    get:
      operationId: listPersonFaces
      summary: List faces belonging to a person
      tags: [people]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Face list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaceListEnvelope"
        "400":
          description: Invalid person ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/people/{id}/events:
    get:
      operationId: listPersonEvents
      summary: List events where this person was detected
      tags: [people]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonEventListEnvelope"
        "404":
          description: Person not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/people/merge:
    post:
      operationId: mergePeople
      summary: Merge two people into one
      tags: [people]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergePeopleRequest"
      responses:
        "200":
          description: People merged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusMerged"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/faces/unmatched:
    get:
      operationId: listUnmatchedFaces
      summary: List faces not assigned to any person
      tags: [faces]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Unmatched face list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaceListEnvelope"
  /api/faces/{id}/assign:
    put:
      operationId: assignFace
      summary: Assign a face to a person
      tags: [faces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignFaceRequest"
      responses:
        "200":
          description: Face assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignFaceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/faces/{id}/crop:
    get:
      operationId: getFaceCrop
      summary: Get cropped face image
      tags: [faces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: JPEG face crop
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: No crop available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/faces/{id}/ignore:
    post:
      operationId: ignoreFace
      summary: Delete/ignore a face
      tags: [faces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Face ignored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusIgnored"
        "400":
          description: Invalid face ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    StatusDeleted:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [deleted]
    StatusUpdated:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [updated]
    StatusMerged:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [merged]
    StatusIgnored:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ignored]
...
    PersonDetailFace:
      type: object
      required: [id, camera, confidence, crop_path, timestamp]
      properties:
        id:
          type: integer
          format: int64
        camera:
          type: string
        confidence:
          type: number
          format: double
        similarity:
          type: number
          format: double
        crop_path:
          type: string
        timestamp:
          type: string
          format: date-time
    PersonDetail:
      type: object
      required: [id, name, ignore, face_count, created_at, faces]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        ignore:
          type: boolean
        face_count:
          type: integer
        created_at:
          type: string
          format: date-time
        faces:
          type: array
          items:
            $ref: "#/components/schemas/PersonDetailFace"
    Face:
      type: object
      required: [id, event_id, camera, confidence, timestamp, created_at]
      properties:
        id:
          type: integer
          format: int64
        event_id:
          type: string
        camera:
          type: string
        person_id:
          type: integer
          format: int64
        crop_path:
          type: string
        confidence:
          type: number
          format: double
        similarity:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    UpdatePersonRequest:
      type: object
      properties:
        name:
          type: string
        ignore:
          type: boolean
    MergePeopleRequest:
      type: object
      required: [source_id, target_id]
      properties:
        source_id:
          type: integer
          format: int64
        target_id:
          type: integer
          format: int64
    AssignFaceRequest:
      type: object
      properties:
        person_id:
          type: integer
          format: int64
        name:
          type: string
    AssignFaceResponse:
      type: object
      required: [status, person_id]
      properties:
        status:
          type: string
          enum: [assigned]
        person_id:
          type: integer
          format: int64
...
    Event:
      type: object
      required: [id, camera, label, score, box, timestamp, snapshot_available, clip_available]
      properties:
        id:
          type: string
        camera:
          type: string
        label:
          type: string
        score:
          type: number
          format: float
        box:
          type: array
          items:
            type: integer
          minItems: 4
          maxItems: 4
        timestamp:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        snapshot_path:
          type: string
        snapshot_available:
          type: boolean
        clip_path:
          type: string
        clip_available:
          type: boolean
        zone_name:
          type: string
        object_name:
          type: string
        sub_label:
          type: string
        category:
          type: string
          enum: [alert, detection]
          description: Review/notification priority. "detection" is low priority (e.g. a stationary/parked vehicle); everything else is "alert".
        kind:
          type: string
        answered_at:
          type: string
          format: date-time
        answered_by:
          type: string
...
    FaceListEnvelope:
      type: object
      required: [items, total, limit, offset, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Face"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    PersonEventListEnvelope:
      type: object
      required: [items, total, limit, offset, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
...